+++ src/include/janetconf.h
@@ -25,7 +25,11 @@
 #ifndef JANETCONF_H
 #define JANETCONF_H
 
-#define JANET_VERSION "0.6.0"
+#define JANET_VERSION_MAJOR 0
+#define JANET_VERSION_MINOR 6
+#define JANET_VERSION_PATCH 0
+#define JANET_VERSION_EXTRA "-dev"
+#define JANET_VERSION "0.6.0-dev"
 
 /* #define JANET_BUILD "local" */